pyproject.toml
[project]
name = "dsai-413-multimodal-rag"
version = "0.1.0"
description = "Multi-Modal RAG QA System for DSAI 413"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Student" }]

dependencies = [
    # Core
    "pdf2image>=1.16.3",
    "pillow>=10.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
    # Retrieval - PyTorch stack for ColVision + bge-small
    "qdrant-client>=1.7.0",
    "sentence-transformers>=2.2.0",
    "transformers>=4.45.0",
    "torch>=2.1.0",
    "accelerate>=0.25.0",
    "colpali-engine>=0.3.10",
    # Parsing - Docling for table structure + layout, pdf2image for rasterization
    "docling>=2.5.0",
    # Generation (Gemini 2.5 Flash)
    "google-genai>=1.0.0",
    # UI
    "streamlit>=1.29.0",
    # Orchestration
    "langgraph>=0.2.0",
    # Eval
    "datasets>=2.14.0",
    "matplotlib>=3.7.0",  # for eval report charts (Hit@k bar chart)
    # Utilities
    "tqdm>=4.66.0",
    "requests>=2.31.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "ipykernel>=6.27.0",
    "reportlab>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu121" }
torchvision = { index = "pytorch-cu121" }

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
exclude = ["tests/", "data/"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
asyncio_mode = "auto"
